+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/EntityTableExportPanelTest.java
@@ -50,7 +50,8 @@ public final class EntityTableExportPanelTest {
 		EntityTablePanel tablePanel = new EntityTablePanel(tableModel, config -> config.includeExport(true));
 		EntityTableExportPanel exportPanel = tablePanel.exportPanel();
 
-		exportPanel.model().treeModel().includeDefault();
+		exportPanel.model().treeModel().includeNone();
+		exportPanel.model().treeModel().includeAll();
 
 		// Save preferences - should be empty JSON since it matches defaults
 		EntityTablePanelPreferences defaultPreferences = new EntityTablePanelPreferences(tablePanel);